+++ library/hls/src/test/java/com/google/android/exoplayer2/source/hls/HlsMediaSourceTest.java
@@ -21,6 +21,7 @@ import static com.google.common.truth.Truth.assertThat;
 import android.net.Uri;
 import android.os.SystemClock;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.MediaItem;
 import com.google.android.exoplayer2.ParserException;
 import com.google.android.exoplayer2.Timeline;
@@ -77,6 +78,8 @@ public class HlsMediaSourceTest {
     // The target live offset is picked from target duration (3 * 4 = 12 seconds) and then expressed
     // in relation to the live edge (12 + 1 seconds).
     assertThat(window.liveConfiguration.targetOffsetMs).isEqualTo(13000);
+    assertThat(window.liveConfiguration.minPlaybackSpeed).isEqualTo(1f);
+    assertThat(window.liveConfiguration.maxPlaybackSpeed).isEqualTo(1f);
     assertThat(window.defaultPositionUs).isEqualTo(4000000);
   }
 
@@ -100,7 +103,7 @@ public class HlsMediaSourceTest {
             + "#EXTINF:4.00000,\n"
             + "fileSequence3.ts\n"
             + "#EXT-X-SERVER-CONTROL:HOLD-BACK=12";
-    // The playlist finishes 1 second before the the current time, therefore there's a live edge
+    // The playlist finishes 1 second before the current time, therefore there's a live edge
     // offset of 1 second.
     SystemClock.setCurrentTimeMillis(Util.parseXsDateTime("2020-01-01T00:00:17.0+00:00"));
     HlsMediaSource.Factory factory = createHlsMediaSourceFactory(playlistUri, playlist);
@@ -113,6 +116,8 @@ public class HlsMediaSourceTest {
     // The target live offset is picked from hold back and then expressed in relation to the live
     // edge (+1 seconds).
     assertThat(window.liveConfiguration.targetOffsetMs).isEqualTo(13000);
+    assertThat(window.liveConfiguration.minPlaybackSpeed).isEqualTo(C.RATE_UNSET);
+    assertThat(window.liveConfiguration.maxPlaybackSpeed).isEqualTo(C.RATE_UNSET);
     assertThat(window.defaultPositionUs).isEqualTo(4000000);
   }
 
@@ -139,7 +144,7 @@ public class HlsMediaSourceTest {
             + "#EXTINF:4.00000,\n"
             + "fileSequence3.ts\n"
             + "#EXT-X-SERVER-CONTROL:HOLD-BACK=12,PART-HOLD-BACK=3";
-    // The playlist finishes 1 second before the the current time.
+    // The playlist finishes 1 second before the current time.
     SystemClock.setCurrentTimeMillis(Util.parseXsDateTime("2020-01-01T00:00:17.0+00:00"));
     HlsMediaSource.Factory factory = createHlsMediaSourceFactory(playlistUri, playlist);
     MediaItem mediaItem = MediaItem.fromUri(playlistUri);
@@ -151,6 +156,8 @@ public class HlsMediaSourceTest {
     // The target live offset is picked from hold back and then expressed in relation to the live
     // edge (+1 seconds).
     assertThat(window.liveConfiguration.targetOffsetMs).isEqualTo(13000);
+    assertThat(window.liveConfiguration.minPlaybackSpeed).isEqualTo(C.RATE_UNSET);
+    assertThat(window.liveConfiguration.maxPlaybackSpeed).isEqualTo(C.RATE_UNSET);
     assertThat(window.defaultPositionUs).isEqualTo(4000000);
   }
 
@@ -182,6 +189,8 @@ public class HlsMediaSourceTest {
     // The target live offset is picked from part hold back and then expressed in relation to the
     // live edge (+1 seconds).
     assertThat(window.liveConfiguration.targetOffsetMs).isEqualTo(4000);
+    assertThat(window.liveConfiguration.minPlaybackSpeed).isEqualTo(C.RATE_UNSET);
+    assertThat(window.liveConfiguration.maxPlaybackSpeed).isEqualTo(C.RATE_UNSET);
     assertThat(window.defaultPositionUs).isEqualTo(0);
   }
 
@@ -395,7 +404,7 @@ public class HlsMediaSourceTest {
             + "#EXTINF:4.00000,\n"
             + "fileSequence0.ts\n"
             + "#EXT-X-SERVER-CONTROL:HOLD-BACK=12,PART-HOLD-BACK=3";
-    // The playlist finishes 1 second before the the current time. This should not affect the target
+    // The playlist finishes 1 second before the current time. This should not affect the target
     // live offset set in the media item.
     SystemClock.setCurrentTimeMillis(Util.parseXsDateTime("2020-01-01T00:00:05.0+00:00"));
     HlsMediaSource.Factory factory = createHlsMediaSourceFactory(playlistUri, playlist);
@@ -411,6 +420,76 @@ public class HlsMediaSourceTest {
 
     Timeline.Window window = timeline.getWindow(0, new Timeline.Window());
     // The target live offset is picked from the media item and not adjusted.
+    assertThat(window.liveConfiguration).isEqualTo(mediaItem.liveConfiguration);
+    assertThat(window.defaultPositionUs).isEqualTo(0);
+  }
+
+  @Test
+  public void loadLivePlaylist_noHoldBackInPlaylistAndNoPlaybackSpeedInMediaItem_usesUnitSpeed()
+      throws TimeoutException, ParserException {
+    String playlistUri = "fake://foo.bar/media0/playlist.m3u8";
+    // The playlist has no hold back defined.
+    String playlist =
+        "#EXTM3U\n"
+            + "#EXT-X-PROGRAM-DATE-TIME:2020-01-01T00:00:00.0+00:00\n"
+            + "#EXT-X-TARGETDURATION:4\n"
+            + "#EXT-X-VERSION:3\n"
+            + "#EXT-X-MEDIA-SEQUENCE:0\n"
+            + "#EXTINF:4.00000,\n"
+            + "fileSequence0.ts";
+    // The playlist finishes 1 second before the current time. This should not affect the target
+    // live offset set in the media item.
+    SystemClock.setCurrentTimeMillis(Util.parseXsDateTime("2020-01-01T00:00:05.0+00:00"));
+    HlsMediaSource.Factory factory = createHlsMediaSourceFactory(playlistUri, playlist);
+    MediaItem mediaItem =
+        new MediaItem.Builder()
+            .setUri(playlistUri)
+            .setLiveConfiguration(
+                new MediaItem.LiveConfiguration.Builder().setTargetOffsetMs(1000).build())
+            .build();
+    HlsMediaSource mediaSource = factory.createMediaSource(mediaItem);
+
+    Timeline timeline = prepareAndWaitForTimeline(mediaSource);
+
+    Timeline.Window window = timeline.getWindow(0, new Timeline.Window());
+    assertThat(window.liveConfiguration.minPlaybackSpeed).isEqualTo(1f);
+    assertThat(window.liveConfiguration.maxPlaybackSpeed).isEqualTo(1f);
+    assertThat(window.defaultPositionUs).isEqualTo(0);
+  }
+
+  @Test
+  public void
+      loadLivePlaylist_noHoldBackInPlaylistAndPlaybackSpeedInMediaItem_usesMediaItemConfiguration()
+          throws TimeoutException, ParserException {
+    String playlistUri = "fake://foo.bar/media0/playlist.m3u8";
+    // The playlist has no hold back defined.
+    String playlist =
+        "#EXTM3U\n"
+            + "#EXT-X-PROGRAM-DATE-TIME:2020-01-01T00:00:00.0+00:00\n"
+            + "#EXT-X-TARGETDURATION:4\n"
+            + "#EXT-X-VERSION:3\n"
+            + "#EXT-X-MEDIA-SEQUENCE:0\n"
+            + "#EXTINF:4.00000,\n"
+            + "fileSequence0.ts";
+    // The playlist finishes 1 second before the current time. This should not affect the target
+    // live offset set in the media item.
+    SystemClock.setCurrentTimeMillis(Util.parseXsDateTime("2020-01-01T00:00:05.0+00:00"));
+    HlsMediaSource.Factory factory = createHlsMediaSourceFactory(playlistUri, playlist);
+    MediaItem mediaItem =
+        new MediaItem.Builder()
+            .setUri(playlistUri)
+            .setLiveConfiguration(
+                new MediaItem.LiveConfiguration.Builder()
+                    .setTargetOffsetMs(1000)
+                    .setMinPlaybackSpeed(0.94f)
+                    .setMaxPlaybackSpeed(1.02f)
+                    .build())
+            .build();
+    HlsMediaSource mediaSource = factory.createMediaSource(mediaItem);
+
+    Timeline timeline = prepareAndWaitForTimeline(mediaSource);
+
+    Timeline.Window window = timeline.getWindow(0, new Timeline.Window());
     assertThat(window.liveConfiguration).isEqualTo(mediaItem.liveConfiguration);
     assertThat(window.defaultPositionUs).isEqualTo(0);
   }